add user and facility-user attributes as tab to member attributes page * new component one-entity-attribute-page.component added for page where we are viewing attributes of one entity
