+++ apps/admin-gui/src/app/facilities/pages/facility-detail-page/facility-attributes/facility-attributes.component.ts
@@ -1,22 +1,7 @@
-import {Component, HostBinding, OnInit, ViewChild} from '@angular/core';
+import {Component, HostBinding, OnInit} from '@angular/core';
 import {ActivatedRoute} from '@angular/router';
-import { MatDialog } from '@angular/material/dialog';
-import { GuiAuthResolver, NotificatorService } from '@perun-web-apps/perun/services';
-import {TranslateService} from '@ngx-translate/core';
-import {AttributesListComponent} from '@perun-web-apps/perun/components';
-import {SelectionModel} from '@angular/cdk/collections';
-import {
-  DeleteAttributeDialogComponent
-} from '../../../../shared/components/dialogs/delete-attribute-dialog/delete-attribute-dialog.component';
-import { getDefaultDialogConfig } from '@perun-web-apps/perun/utils';
-import { Attribute, AttributesManagerService, Facility } from '@perun-web-apps/perun/openapi';
-import { PageEvent } from '@angular/material/paginator';
-import {
-  TABLE_ATTRIBUTES_SETTINGS,
-  TableConfigService
-} from '@perun-web-apps/config/table-config';
-import { EditAttributeDialogComponent } from '@perun-web-apps/perun/dialogs';
-import { CreateAttributeDialogComponent } from '../../../../shared/components/dialogs/create-attribute-dialog/create-attribute-dialog.component';
+import { GuiAuthResolver } from '@perun-web-apps/perun/services';
+import { Facility } from '@perun-web-apps/perun/openapi';
 
 @Component({
   selector: 'app-facility-attributes',
@@ -27,35 +12,13 @@ export class FacilityAttributesComponent implements OnInit {
 
   @HostBinding('class.router-component') true;
 
-  constructor(private attributesManager: AttributesManagerService,
-              private route: ActivatedRoute,
-              private dialog: MatDialog,
-              private notificator: NotificatorService,
-              private tableConfigService: TableConfigService,
-              private translate: TranslateService,
-              private authResolver: GuiAuthResolver) {
-    this.translate.get('FACILITY_DETAIL.SETTINGS.ATTRIBUTES.SUCCESS_SAVE').subscribe(value => this.saveSuccessMessage = value);
-    this.translate.get('FACILITY_DETAIL.SETTINGS.ATTRIBUTES.SUCCESS_DELETE').subscribe(value => this.deleteSuccessMessage = value);
-  }
+  constructor(private route: ActivatedRoute,
+              private authResolver: GuiAuthResolver) {}
 
-  @ViewChild('list')
-  list: AttributesListComponent;
-
-  attributes: Attribute[] = [];
-  selection = new SelectionModel<Attribute>(true, []);
   facilityId: number;
-  saveSuccessMessage: string;
-  deleteSuccessMessage: string;
-
   facilityUserAttAuth: boolean;
 
-  loading: boolean;
-  filterValue = '';
-  pageSize: number;
-  tableId = TABLE_ATTRIBUTES_SETTINGS;
-
   ngOnInit() {
-    this.pageSize = this.tableConfigService.getTablePageSize(this.tableId);
     this.route.parent.params.subscribe(params => {
       this.facilityId = params['facilityId'];
 
@@ -65,84 +28,6 @@ export class FacilityAttributesComponent implements OnInit {
       }
       this.facilityUserAttAuth = this.authResolver.isAuthorized('getAssignedUsers_Facility_policy', [facility]);
 
-      this.refreshTable();
     });
-  }
-
-  onDelete() {
-    const config = getDefaultDialogConfig();
-    config.width = '450px';
-    config.data = {
-      entityId: this.facilityId,
-      entity: 'facility',
-      attributes: this.selection.selected,
-      theme: 'facility-theme'
-    };
-
-    const dialogRef = this.dialog.open(DeleteAttributeDialogComponent, config);
-
-    dialogRef.afterClosed().subscribe(result => {
-      if (result) {
-        this.refreshTable();
-      }
-    });
-  }
-
-  onSave() {
-    // have to use this to update attribute with map in it, before saving it
-    this.list.updateMapAttributes();
-
-    const config = getDefaultDialogConfig();
-    config.width = '450px';
-    config.data = {
-      entityId: this.facilityId,
-      entity: 'facility',
-      attributes: this.selection.selected
-    };
-
-    const dialogRef = this.dialog.open(EditAttributeDialogComponent, config);
-
-    dialogRef.afterClosed().subscribe(result => {
-      if (result) {
-        this.refreshTable();
-      }
-    });
-  }
-
-  onCreate() {
-    const config = getDefaultDialogConfig();
-    config.width = '1050px';
-    config.data = {
-      entityId: this.facilityId,
-      entity: 'facility',
-      notEmptyAttributes: this.attributes,
-      style: 'facility-theme'
-    };
-
-    const dialogRef = this.dialog.open(CreateAttributeDialogComponent, config);
-
-    dialogRef.afterClosed().subscribe(result => {
-      if (result === 'saved') {
-        this.refreshTable();
-      }
-    });
-  }
-
-  refreshTable() {
-    this.loading = true;
-    this.attributesManager.getFacilityAttributes(this.facilityId).subscribe(attributes => {
-      this.attributes = attributes;
-      this.selection.clear();
-      this.loading = false;
-    });
-  }
-
-  applyFilter(filterValue: string) {
-    this.filterValue = filterValue;
-  }
-
-  pageChanged(event: PageEvent) {
-    this.pageSize = event.pageSize;
-    this.tableConfigService.setTablePageSize(this.tableId, event.pageSize);
   }
 }
