+++ apps/admin-gui/src/app/shared/components/one-entity-attribute-page/one-entity-attribute-page.component.html
@@ -1,0 +1,26 @@
+<perun-web-apps-refresh-button class="mt-2" (refresh)="refreshTable()"></perun-web-apps-refresh-button>
+<button mat-flat-button class="mr-2 mt-2" color="accent" (click)="onCreate()">
+  {{'USER_DETAIL.SETTINGS.ATTRIBUTES.CREATE' | translate}}
+</button>
+
+<button mat-flat-button class="mr-2 mt-2" color="accent" (click)="onSave()" [disabled]="selection.selected.length === 0">
+  {{'USER_DETAIL.SETTINGS.ATTRIBUTES.SAVE' | translate}}
+</button>
+
+<button mat-flat-button class="mt-2" color="warn" (click)="onDelete()" [disabled]="selection.selected.length === 0">
+  {{'USER_DETAIL.SETTINGS.ATTRIBUTES.DELETE' | translate}}
+</button>
+<perun-web-apps-immediate-filter (filter)="applyFilter($event)"
+                                 [placeholder]="'SHARED_LIB.PERUN.COMPONENTS.ATTRIBUTES_LIST.FILTER'"></perun-web-apps-immediate-filter>
+
+<mat-spinner class="ml-auto mr-auto" *ngIf="loading"></mat-spinner>
+
+<perun-web-apps-attributes-list
+  (page)="pageChanged($event)"
+  [pageSize]="pageSize"
+  #list
+  *ngIf="!loading"
+  [filterValue]="filterValue"
+  [attributes]="attributes"
+  [selection]="selection">
+</perun-web-apps-attributes-list>
