+++ apps/admin-gui/src/app/users/pages/user-detail-page/user-attributes/user-attributes.component.html
@@ -4,32 +4,7 @@
     <ng-template matTabLabel>
       {{'ADMIN_USER.ENTITY' | translate}}
     </ng-template>
-    <perun-web-apps-refresh-button class="mt-2" (refresh)="refreshTable()"></perun-web-apps-refresh-button>
-    <button mat-flat-button class="mr-2 mt-2" color="accent" (click)="onCreate()">
-      {{'USER_DETAIL.SETTINGS.ATTRIBUTES.CREATE' | translate}}
-    </button>
-
-    <button mat-flat-button class="mr-2 mt-2" color="accent" (click)="onSave()" [disabled]="selection.selected.length === 0">
-      {{'USER_DETAIL.SETTINGS.ATTRIBUTES.SAVE' | translate}}
-    </button>
-
-    <button mat-flat-button class="mt-2" color="warn" (click)="onDelete()" [disabled]="selection.selected.length === 0">
-      {{'USER_DETAIL.SETTINGS.ATTRIBUTES.DELETE' | translate}}
-    </button>
-    <perun-web-apps-immediate-filter (filter)="applyFilter($event)"
-                                     [placeholder]="'SHARED_LIB.PERUN.COMPONENTS.ATTRIBUTES_LIST.FILTER'"></perun-web-apps-immediate-filter>
-
-    <mat-spinner class="ml-auto mr-auto" *ngIf="loading"></mat-spinner>
-
-    <perun-web-apps-attributes-list
-      (page)="pageChanged($event)"
-      [pageSize]="pageSize"
-      #list
-      *ngIf="!loading"
-      [filterValue]="filterValue"
-      [attributes]="attributes"
-      [selection]="selection">
-    </perun-web-apps-attributes-list>
+    <app-one-entity-attribute-page [entityId]="userId" [entity]="'user'"></app-one-entity-attribute-page>
   </mat-tab>
   <mat-tab *ngIf="userFacilityAttAuth">
     <ng-template matTabLabel>
@@ -39,7 +14,7 @@
     </ng-template>
     <ng-template matTabContent>
       <app-two-entity-attribute-page [firstEntity]="'user'" [firstEntityId]="userId"
-                                     [secondEntity]="'facility'" *ngIf="!loading"></app-two-entity-attribute-page>
+                                     [secondEntity]="'facility'"></app-two-entity-attribute-page>
     </ng-template>
   </mat-tab>
 </mat-tab-group>
