+++ apps/admin-gui/src/app/vos/pages/group-detail-page/group-attributes/group-attributes.component.html
@@ -5,31 +5,7 @@
       <ng-template matTabLabel>
         {{'GROUP_DETAIL.ENTITY' | translate}}
       </ng-template>
-      <perun-web-apps-refresh-button class="mt-2" (refresh)="refreshTable()"></perun-web-apps-refresh-button>
-      <button mat-flat-button class="mr-2 mt-2" color="accent" (click)="onCreate()">
-        {{'GROUP_DETAIL.SETTINGS.ATTRIBUTES.CREATE' | translate}}
-      </button>
-
-      <button mat-flat-button class="mr-2 mt-2" color="accent" (click)="onSave()" [disabled]="selection.selected.length === 0">
-        {{'GROUP_DETAIL.SETTINGS.ATTRIBUTES.SAVE' | translate}}
-      </button>
-
-      <button mat-flat-button class="mt-2" color="warn" (click)="onDelete()" [disabled]="selection.selected.length === 0">
-        {{'GROUP_DETAIL.SETTINGS.ATTRIBUTES.DELETE' | translate}}
-      </button>
-      <perun-web-apps-immediate-filter (filter)="applyFilter($event)" [placeholder]="'SHARED_LIB.PERUN.COMPONENTS.ATTRIBUTES_LIST.FILTER'"></perun-web-apps-immediate-filter>
-
-      <mat-spinner class="ml-auto mr-auto" *ngIf="loading"></mat-spinner>
-
-      <perun-web-apps-attributes-list
-        (page)="pageChanged($event)"
-        [pageSize]="pageSize"
-        #list
-        *ngIf="!loading"
-        [filterValue]="filterValue"
-        [attributes]="attributes"
-        [selection]="selection">
-      </perun-web-apps-attributes-list>
+      <app-one-entity-attribute-page [entity]="'group'" [entityId]="groupId"></app-one-entity-attribute-page>
     </mat-tab>
     <mat-tab *ngIf="groupMemberAttAuth">
       <ng-template matTabLabel>
@@ -39,7 +15,7 @@
       </ng-template>
       <ng-template matTabContent>
         <app-two-entity-attribute-page [firstEntity]="'group'" [firstEntityId]="groupId"
-                                       [secondEntity]="'member'" *ngIf="!loading"></app-two-entity-attribute-page>
+                                       [secondEntity]="'member'"></app-two-entity-attribute-page>
       </ng-template>
     </mat-tab>
     <mat-tab *ngIf="groupResourceAttAuth">
@@ -50,7 +26,7 @@
       </ng-template>
       <ng-template matTabContent>
         <app-two-entity-attribute-page [firstEntity]="'group'" [firstEntityId]="groupId"
-                                       [secondEntity]="'resource'" *ngIf="!loading"></app-two-entity-attribute-page>
+                                       [secondEntity]="'resource'"></app-two-entity-attribute-page>
       </ng-template>
     </mat-tab>
   </mat-tab-group>
