+++ apps/admin-gui/src/app/vos/pages/member-detail-page/member-attributes/member-attributes.component.ts
@@ -1,27 +1,10 @@
-import {Component, HostBinding, OnInit, ViewChild} from '@angular/core';
+import {Component, HostBinding, OnInit} from '@angular/core';
 import {ActivatedRoute} from '@angular/router';
-import { GuiAuthResolver, NotificatorService } from '@perun-web-apps/perun/services';
-import { MatDialog } from '@angular/material/dialog';
-import {TranslateService} from '@ngx-translate/core';
-import {AttributesListComponent} from '@perun-web-apps/perun/components';
-import {SelectionModel} from '@angular/cdk/collections';
+import { GuiAuthResolver } from '@perun-web-apps/perun/services';
 import {
-  DeleteAttributeDialogComponent
-} from '../../../../shared/components/dialogs/delete-attribute-dialog/delete-attribute-dialog.component';
-import { getDefaultDialogConfig } from '@perun-web-apps/perun/utils';
-import {
-  Attribute,
-  AttributesManagerService, Group,
-  GroupsManagerService, Member, MembersManagerService, Resource,
-  ResourcesManagerService
+  Member,
+  MembersManagerService,
 } from '@perun-web-apps/perun/openapi';
-import {
-  TABLE_ATTRIBUTES_SETTINGS,
-  TableConfigService
-} from '@perun-web-apps/config/table-config';
-import { PageEvent } from '@angular/material/paginator';
-import { EditAttributeDialogComponent } from '@perun-web-apps/perun/dialogs';
-import { CreateAttributeDialogComponent } from '../../../../shared/components/dialogs/create-attribute-dialog/create-attribute-dialog.component';
 
 @Component({
   selector: 'app-member-attributes',
@@ -33,38 +16,18 @@ export class MemberAttributesComponent implements OnInit {
   @HostBinding('class.router-component') true;
   constructor(
     private route: ActivatedRoute,
-    private attributesManager: AttributesManagerService,
-    private notificator: NotificatorService,
-    private dialog: MatDialog,
-    private translate: TranslateService,
-    private tableConfigService: TableConfigService,
     private authResolver: GuiAuthResolver,
     private memberManager: MembersManagerService
-  ) {
-    this.translate.get('MEMBER_DETAIL.SETTINGS.ATTRIBUTES.SUCCESS_SAVE').subscribe(value => this.saveSuccessMessage = value);
-    this.translate.get('MEMBER_DETAIL.SETTINGS.ATTRIBUTES.SUCCESS_DELETE').subscribe(value => this.deleteSuccessMessage = value);
-  }
+  ) {}
 
-  @ViewChild('list')
-  list: AttributesListComponent;
-
-  saveSuccessMessage: string;
-  deleteSuccessMessage: string;
-  selection = new SelectionModel<Attribute>(true, []);
-  attributes: Attribute[] = [];
   memberId: number;
   member: Member;
 
   memberResourceAttAuth: boolean;
   memberGroupAttAuth: boolean;
-
-  loading: boolean;
-  filterValue = '';
-  tableId = TABLE_ATTRIBUTES_SETTINGS;
-  pageSize: number;
+  userFacilityAttAuth: boolean;
 
   ngOnInit() {
-    this.pageSize = this.tableConfigService.getTablePageSize(this.tableId);
     this.route.parent.params.subscribe(params => {
       this.memberId = params['memberId'];
 
@@ -73,84 +36,9 @@ export class MemberAttributesComponent implements OnInit {
 
        this.memberGroupAttAuth = this.authResolver.isAuthorized('getMemberGroups_Member_policy', [this.member]);
        this.memberResourceAttAuth = this.authResolver.isAuthorized('getAllowedResources_Member_policy', [this.member]);
-       this.refreshTable();
+       this.userFacilityAttAuth = this.authResolver.isAuthorized('getAssignedFacilities_User_policy', [{beanName: 'User', id: member.userId}]);
      });
     });
   }
 
-  onCreate() {
-    const config = getDefaultDialogConfig();
-    config.width = '1050px';
-    config.data = {
-      entityId: this.memberId,
-      entity: 'member',
-      notEmptyAttributes: this.attributes,
-      style: 'member-theme'
-    };
-
-    const dialogRef = this.dialog.open(CreateAttributeDialogComponent, config);
-
-    dialogRef.afterClosed().subscribe(result => {
-      if (result) {
-        this.refreshTable();
-      }
-    });
-  }
-
-  onSave() {
-    // have to use this to update attribute with map in it, before saving it
-    this.list.updateMapAttributes();
-
-    const config = getDefaultDialogConfig();
-    config.width = '450px';
-    config.data = {
-      entityId: this.memberId,
-      entity: 'member',
-      attributes: this.selection.selected
-    };
-
-    const dialogRef = this.dialog.open(EditAttributeDialogComponent, config);
-
-    dialogRef.afterClosed().subscribe(result => {
-      if (result) {
-        this.refreshTable();
-      }
-    });
-  }
-
-  onDelete() {
-    const config = getDefaultDialogConfig();
-    config.width = '450px';
-    config.data = {
-      entityId: this.memberId,
-      entity: 'member',
-      attributes: this.selection.selected
-    };
-
-    const dialogRef = this.dialog.open(DeleteAttributeDialogComponent, config);
-
-    dialogRef.afterClosed().subscribe(didConfirm => {
-      if (didConfirm) {
-        this.refreshTable();
-      }
-    });
-  }
-
-  refreshTable() {
-    this.loading = true;
-    this.attributesManager.getMemberAttributes(this.memberId).subscribe(attributes => {
-      this.attributes = attributes;
-      this.selection.clear();
-      this.loading = false;
-    });
-  }
-
-  applyFilter(filterValue: string) {
-    this.filterValue = filterValue;
-  }
-
-  pageChanged(event: PageEvent) {
-    this.pageSize = event.pageSize;
-    this.tableConfigService.setTablePageSize(this.tableId, event.pageSize);
-  }
 }
