+++ apps/admin-gui/src/app/vos/pages/vo-detail-page/vo-attributes/vo-attributes.component.html
@@ -1,42 +1,4 @@
 <div>
   <h1 class="page-subtitle">{{'VO_DETAIL.SETTINGS.ATTRIBUTES.TITLE' | translate}}</h1>
-  <perun-web-apps-refresh-button (refresh)="refreshTable()"></perun-web-apps-refresh-button>
-  <button
-    *ngIf="!loading"
-    mat-flat-button
-    class="mr-2"
-    color="accent"
-    (click)="onCreate()">
-    {{'VO_DETAIL.SETTINGS.ATTRIBUTES.CREATE' | translate}}
-  </button>
-
-  <button
-    *ngIf="!loading"
-    mat-flat-button
-    class="mr-2"
-    color="accent"
-    (click)="onSave()"
-    [disabled]="selection.selected.length === 0">
-    {{'VO_DETAIL.SETTINGS.ATTRIBUTES.SAVE' | translate}}
-  </button>
-
-  <button
-    *ngIf="!loading"
-    mat-flat-button
-    color="warn"
-    (click)="onDelete()"
-    [disabled]="selection.selected.length === 0">
-    {{'VO_DETAIL.SETTINGS.ATTRIBUTES.DELETE' | translate}}
-  </button>
-  <perun-web-apps-immediate-filter (filter)="applyFilter($event)" [placeholder]="'SHARED_LIB.PERUN.COMPONENTS.ATTRIBUTES_LIST.FILTER'"></perun-web-apps-immediate-filter>
-  <mat-spinner class="ml-auto mr-auto" *ngIf="loading"></mat-spinner>
-  <perun-web-apps-attributes-list
-    (page)="pageChanged($event)"
-    [pageSize]="pageSize"
-    #list
-    *ngIf="!loading"
-    [filterValue]="filterValue"
-    [attributes]="attributes"
-    [selection]="selection">
-  </perun-web-apps-attributes-list>
+  <app-one-entity-attribute-page [entityId]="voId" [entity]="'vo'"></app-one-entity-attribute-page>
 </div>
