+++ apps/admin-gui/src/app/vos/pages/vo-detail-page/vo-attributes/vo-attributes.component.ts
@@ -1,22 +1,5 @@
 import {ActivatedRoute} from '@angular/router';
-import {SelectionModel} from '@angular/cdk/collections';
-import {
-  DeleteAttributeDialogComponent
-} from '../../../../shared/components/dialogs/delete-attribute-dialog/delete-attribute-dialog.component';
-import { MatDialog } from '@angular/material/dialog';
-import {Component, HostBinding, OnInit, ViewChild} from '@angular/core';
-import {AttributesListComponent} from '@perun-web-apps/perun/components';
-import {NotificatorService} from '@perun-web-apps/perun/services';
-import {TranslateService} from '@ngx-translate/core';
-import { getDefaultDialogConfig } from '@perun-web-apps/perun/utils';
-import { Attribute, AttributesManagerService } from '@perun-web-apps/perun/openapi';
-import {
-  TABLE_ATTRIBUTES_SETTINGS,
-  TableConfigService
-} from '@perun-web-apps/config/table-config';
-import { PageEvent } from '@angular/material/paginator';
-import { EditAttributeDialogComponent } from '@perun-web-apps/perun/dialogs';
-import { CreateAttributeDialogComponent } from '../../../../shared/components/dialogs/create-attribute-dialog/create-attribute-dialog.component';
+import {Component, HostBinding, OnInit} from '@angular/core';
 
 @Component({
   selector: 'app-vo-attributes',
@@ -27,111 +10,13 @@ export class VoAttributesComponent implements OnInit {
 
   @HostBinding('class.router-component') true;
 
-  constructor(private attributesManager: AttributesManagerService,
-              private route: ActivatedRoute,
-              private dialog: MatDialog,
-              private notificator: NotificatorService,
-              private tableConfigService: TableConfigService,
-              private translate: TranslateService) {
-    this.translate.get('VO_DETAIL.SETTINGS.ATTRIBUTES.SUCCESS_SAVE').subscribe(value => this.saveSuccessMessage = value);
-    this.translate.get('VO_DETAIL.SETTINGS.ATTRIBUTES.SUCCESS_DELETE').subscribe(value => this.deleteSuccessMessage = value);
-  }
+  constructor(private route: ActivatedRoute) { }
 
-  @ViewChild('list')
-  list: AttributesListComponent;
-
-  attributes: Attribute[] = [];
-  selection = new SelectionModel<Attribute>(true, []);
   voId: number;
-  saveSuccessMessage: string;
-  deleteSuccessMessage: string;
-
-  loading: boolean;
-  filterValue = '';
-  tableId = TABLE_ATTRIBUTES_SETTINGS;
-  pageSize: number;
 
   ngOnInit() {
-    this.pageSize = this.tableConfigService.getTablePageSize(this.tableId);
     this.route.parent.params.subscribe(parentParams => {
       this.voId = parentParams['voId'];
-      this.refreshTable();
     });
-  }
-
-  onDelete() {
-    const config = getDefaultDialogConfig();
-    config.width = '450px';
-    config.data = {
-      entityId: this.voId,
-      entity: 'vo',
-      attributes: this.selection.selected
-    };
-
-    const dialogRef = this.dialog.open(DeleteAttributeDialogComponent, config);
-
-    dialogRef.afterClosed().subscribe(result => {
-      if (result) {
-        this.refreshTable();
-      }
-    });
-  }
-
-  onCreate() {
-    const config = getDefaultDialogConfig();
-    config.width = '1050px';
-    config.data = {
-      entityId: this.voId,
-      entity: 'vo',
-      notEmptyAttributes: this.attributes,
-      style: 'vo-theme'
-    };
-
-    const dialogRef = this.dialog.open(CreateAttributeDialogComponent, config);
-
-    dialogRef.afterClosed().subscribe(result => {
-      if (result === 'saved') {
-        this.refreshTable();
-      }
-    });
-  }
-
-  onSave() {
-    // have to use this to update attribute with map in it, before saving it
-    this.list.updateMapAttributes();
-
-    const config = getDefaultDialogConfig();
-    config.width = '450px';
-    config.data = {
-      entityId: this.voId,
-      entity: 'vo',
-      attributes: this.selection.selected
-    };
-
-    const dialogRef = this.dialog.open(EditAttributeDialogComponent, config);
-
-    dialogRef.afterClosed().subscribe(result => {
-      if (result) {
-        this.refreshTable();
-      }
-    });
-  }
-
-  refreshTable() {
-    this.loading = true;
-    this.attributesManager.getVoAttributes(this.voId).subscribe(attributes => {
-      this.attributes = attributes;
-      this.selection.clear();
-      this.loading = false;
-    });
-  }
-
-  applyFilter(filterValue: string) {
-    this.filterValue = filterValue;
-  }
-
-  pageChanged(event: PageEvent) {
-    this.pageSize = event.pageSize;
-    this.tableConfigService.setTablePageSize(this.tableId, event.pageSize);
   }
 }
